refactor: transform module exporter to package
The resources export will require family of related functions (similar
to corosync_conf family). The module exporter.py becoming too long (and
messy). Replace module by package with possibility to split to modules
for related functions.
